Slightly changed how hits work Added knockback on player hit

+++ Assets/Scripts/Player/PlayerStats.cs
@@ -6,11 +6,11 @@ public class PlayerStats
     public int MaxHealth;
     public float MaxSpeed;
     public float RangedDelay;
-    public float RangedDamage;
+    public int RangedDamage;
     public float RangedRange;
     public float RangedSpeed;
     public float MeleeRange;
-    public float MeleeDamage;
+    public int MeleeDamage;
     public bool HasMelee = true;
     public bool HasRanged = true;
 }
